fix to the token/ ui changes to the appointment title/ showing some error messages on the snackbar

+++ src/pages/users/userService.ts
@@ -7,10 +7,10 @@ export const getUsers = async () => {
   try {
     const response = await httpClient.get(`users`);
 
-    return response.data; // Return the fetched users
+    return response.data; 
   } catch (error) {
     console.error(error);
-    throw error; // Rethrow the error to handle it in the component
+    throw error; 
   }
 };
 
@@ -18,9 +18,9 @@ export const createUser = async (userData: UserCreate) => {
   try {
     const response = await httpClient.post(`user`, userData);
 
-    return response.data; // Return the created user
+    return response.data; 
   } catch (error) {
-    throw error; // Rethrow the error to handle it in the component
+    throw error; 
   }
 };
 
@@ -28,11 +28,10 @@ export const getUserById = async (userId: string) => {
   try {
     const response = await httpClient.get(`user/${userId}`);
 
-    // Assuming the response contains the user data
-    return response.data; // Return the fetched user by ID
+    return response.data; 
   } catch (error) {
     console.error(error);
-    throw error; // Rethrow the error to handle it in the component
+    throw error; 
   }
 };
 
@@ -43,22 +42,20 @@ export const updateUser = async (
   try {
     const response = await httpClient.put(`user/${userId}`, updatedUserData);
 
-    // Assuming the response contains the updated user
-    return response.data; // Return the updated user
+    return response.data; 
   } catch (error) {
     console.error(error);
-    throw error; // Rethrow the error to handle it in the component
+    throw error;
   }
 };
 
-export const deleteUser = async (userId: string) => {
+export const deleteUser = async (userId: string,requestingUserId: string) => {
   try {
-    const response = await httpClient.delete(`user/${userId}`);
+    const response = await httpClient.delete(`user/${userId}`, { data: { requestingUserId } });
 
-    // Assuming the response contains a message or confirmation
-    return response.data; // Return the response from the delete operation
+    return response.data; 
   } catch (error) {
     console.error(error);
-    throw error; // Rethrow the error to handle it in the component
+    throw error; 
   }
 };
